Back/TFG/requirements.txt
dj-database-url==2.1.0
Django==5.1.2
django-cors-headers==4.6.0
django-model-utils==5.0.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
wheel==0.38.1
whitenoise==6.6.0
python-dotenv==1.0.1
python-decouple==3.8
pandas
matplotlib
seaborn
pillow
mysqlclient
weasyprint